requirements.txt
streamlit
streamlit_webrtc
opencv-python-headless
av
python-dotenv
openai
pydub
altair==4.2.2